Renamed `mendel-production-middleware` for package name consistency.
