+++ examples/full-example/package.json
@@ -14,7 +14,7 @@
     "react": "^0.14.0",
     "react-dom": "^0.14.0",
     "morgan": "^1.7.0",
-    "mendel-production-middleware": "*"
+    "mendel-middleware": "*"
   },
   "devDependencies": {
     "babel-preset-es2015": "^6.6.0",
