+++ packages/mendel-middleware/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mendel-production-middleware",
+  "name": "mendel-middleware",
   "version": "1.0.0",
   "description": "Mendel Middleware that uses generated manifests to output multilayer resolution of isomorphic applications.",
   "main": "index.js",
